Fix #KUSU-1200: fstab.append is reset to default after reboot of the installer node

+++ src/kits/base/packages/kusu-base-node/sbin/S91KusuHomeFstab.rc.py
@@ -68,6 +68,9 @@ class KusuRC(rcplugin.Plugin):
             # Use CFM append mechanism to append the entry for the home
             # NFS mount
             fstab_append = path('/etc/cfm/%s/etc/fstab.append' % ng.ngname)
+            # Don't clobber the existing file
+            if fstab_append.exists():
+                continue
 
             f = open(fstab_append, 'w')
             f.writelines(entries)
